refactor: 서비스 클래스 => 인터페이스로 변경, 모듈화 수행
