+++ src/main/java/com/itwillbs/factron/controller/sys/SysMainRestController.java
@@ -2,7 +2,7 @@ package com.itwillbs.factron.controller.sys;
 
 import com.itwillbs.factron.dto.ResponseDTO;
 import com.itwillbs.factron.dto.sys.SysMainDTO;
-import com.itwillbs.factron.service.sys.SysMainService;
+import com.itwillbs.factron.service.sys.SysMainServiceImpl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -16,7 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class SysMainRestController {
 
-    private final SysMainService sysMainService;
+    private final SysMainServiceImpl sysMainService;
 
     @PostMapping("")
     public ResponseDTO<Void> saveSysMain(@RequestBody SysMainDTO sysMainDTO) {
